Add comment policy and gates to comment controller

+++ app/Http/Controllers/api/v1/CommentController.php
@@ -9,27 +9,31 @@ use App\Models\Comment;
 use App\Models\Task;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\ResourceCollection;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
     public function index(Task $task): ResourceCollection
     {
+        Gate::authorize('viewAny', [Comment::class, $task]);
         return CommentResource::collection($task->comments()->get());
     }
 
-    public function store(Task $task, StoreRequest $request): Comment
+    public function store(Task $task, StoreRequest $request): CommentResource
     {
+        Gate::authorize('create', [Comment::class, $task]);
         $comment = Comment::query()
             ->create([
                 'user_id' => auth()->user()->id,
                 'task_id' => $task->id,
                 'comment' => $request->comment
             ]);
-        return $comment;
+        return new CommentResource($comment);
     }
 
     public function destroy(Comment $comment): void
     {
+        Gate::authorize('delete', $comment);
         $comment->delete();
     }
 }
